реализован функцтонал

+++ src/components/Card.js
@@ -2,23 +2,32 @@ class Card {
   #cardText;
   #cardImage;
   #cardSelector;
+  #likesCounter;
   #element;
-  #isLiked;
+  #likes;
+  // #isLiked;
   #handleOpenImage;
   #likeButton;
   #deleteButton;
   #placeImage;
   #id;
-  #api
+  #api;
+  #ownerId;
+  #currentUserId;
+  #confirmDeletion;
 
-  constructor(data, cardSelector, handleOpenImage, api) {
+  constructor(data, cardSelector, handleOpenImage, confirmDeletion, api, currentUserId) {
     this.#cardText = data.name;
     this.#cardImage = data.link;
     this.#id = data._id;
+    this.#likes = data.likes;
+    this.#currentUserId = currentUserId;
+    this.#ownerId = data.owner._id
     this.#api = api;
     this.#cardSelector = cardSelector;
-    this.#isLiked = false;
+    // this.#isLiked = false;
     this.#handleOpenImage = handleOpenImage;
+    this.#confirmDeletion = confirmDeletion;
   }
 
   #setEventListeners() {
@@ -35,17 +44,65 @@ class Card {
     });
   }
 
+  //функция проверки наличия лайка текущего пользователя
+  #isLiked() {
+    // this.#likes.forEach(user => {
+    //   return user._id === this.#currentUserId;
+    // })
+    return this.#likes.some(user => user._id === this.#currentUserId);
+    // const likes = this.#likes;
+    // for (let user of likes) {
+    //   if (this.#currentUserId === user._id) {
+    //     return true;
+    //   }
+    // }
+    // return false;
+  }
+
+  #updateLikesControl() {
+    //обновляем количество лайков
+    this.#likesCounter.textContent = this.#likes.length;
+    //отображение счетчика (скрываем при нуле)
+    if (this.#likes.length === 0) { this.#likesCounter.classList.remove('likesCounter_active') }
+    else { this.#likesCounter.classList.add('likesCounter_active') }
+    //отображение лайка на основании того, лайкал ли пользователь карточку
+    if (this.#isLiked()) {
+      this.#likeButton.classList.add('like_active');
+    } else { this.#likeButton.classList.remove('like_active'); }
+  }
+
   #like() {
-    this.#likeButton.classList.toggle('like_active');
-    this.#isLiked = !this.#isLiked;
+    //проверяем есть ли лайк текущего пользователя в массиве
+    if (this.#isLiked()) {
+      this.#api.removeLike(this.#id)
+        .then((card) => {
+          this.#likes = card.likes;
+          this.#updateLikesControl()
+        })
+    } else {
+      this.#api.addLike(this.#id)
+        .then((card) => {
+          this.#likes = card.likes;
+          this.#updateLikesControl()
+        })
+        .catch((err) => {
+          console.log(err); // выведем ошибку в консоль
+        });
+    }
   }
 
   #removeCard() {
-    this.#api.deleteCard(this.#id)
+    this.#confirmDeletion()
+      .then(() => {
+        return this.#api.deleteCard(this.#id);
+      })
       .then(() => {
         this.#element.remove();
         this.#element.null;
       })
+      .catch((err) => {
+        console.log(err); // выведем ошибку в консоль
+      });
   }
 
   #getTemplate() {
@@ -63,9 +120,13 @@ class Card {
     this.#element = this.#getTemplate();
     this.#likeButton = this.#element.querySelector('.like');
     this.#deleteButton = this.#element.querySelector('.trash');
-    this.#placeImage = this.#element.querySelector('.place__image')
+    this.#likesCounter = this.#element.querySelector('.likesCounter');
+    this.#placeImage = this.#element.querySelector('.place__image');
+    if (this.#ownerId != this.#currentUserId) {
+      this.#deleteButton.classList.add('trash_hidden')
+    }
 
-
+    this.#updateLikesControl();
     this.#setEventListeners();
 
     // Добавим данные
